Rename fish routes, pages, and metadata
