+++ asset_client/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "fish_net",
+  "name": "asset_track",
   "version": "0.0.0",
-  "description": "A Sawtooth Supply Chain client web app for trading fish",
+  "description": "A Sawtooth Supply Chain client web app for tracking assets",
   "main": "",
   "scripts": {
     "compile_protos": "node scripts/compile_protos.js > src/protos.json",
