+++ asset_client/public/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>FishNet</title>
+    <title>AssetTrack</title>
   </head>
   <body>
     <!-- mount point for the Mithril JS app -->
