+++ asset_client/src/components/navigation.js
@@ -24,7 +24,7 @@ const m = require('mithril')
 const Navbar = {
   view (vnode) {
     return m('nav.navbar.navbar-expand-sm.navbar-dark.bg-dark.mb-5', [
-      m('a.navbar-brand[href="/"]', { oncreate: m.route.link }, 'FishNet'),
+      m('a.navbar-brand[href="/"]', { oncreate: m.route.link }, 'AssetTrack'),
       m('button.navbar-toggler', {
         type: 'button',
         'data-toggle': 'collapse',
