+++ asset_client/src/views/add_asset_form.js
