+++ asset_client/src/views/asset_detail.js
