+++ asset_client/src/views/list_assets.js
